[REFACTOR] FileExtension 분리
